Move p.emitter data to p.sys, move draw_particles
Moved the ParticleEmitter's particle buffer and related state to
ParticleSystem, so that emitters are pure data oriented objects.
A particle system is responsible for the buffer management of its
emitters and can optimize its layout accordingly.
Furthermore, it's now the GraphicsDevice's responsibility to draw
a particle system, since this gives us some room for optimization.